+++ src/platform/macos_delegate.rs
@@ -23,7 +23,7 @@ fn send_create_window() {
     if let Some(proxy) = PROXY.get() {
         crate::shortcuts::send_app_event(
             proxy,
-            AppEvent::CreateWindow(crate::app::event::WindowRequestOrigin::User),
+            AppEvent::CreateWindow(crate::app::event::WindowRequestOrigin::User, None),
         );
     }
 }